# short-video-service/service/src/main/resources/openapi.yaml
openapi: 3.0.2
info:
  title: Short Video Service
  description: Services to support short videos
  version: "1.0"
paths:
...
  /short-video/comment:
    post:
      tags:
        - Short Video
      summary: Adds a short video comment
      operationId: addComment
      requestBody:
        description: Comment object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShortVideoComment'
        required: true
      responses:
        201:
          description: Created short video comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortVideoStatsResponse'
        400:
          $ref: '#/components/responses/InvalidInput'
        default:
          $ref: '#/components/responses/Error'
    get:
      tags:
        - Short Video
      summary: Retrieves short video comments
      operationId: findComments
      x-spring-paginated: true
      parameters:
        - $ref: '#/components/parameters/videoUuid'
        - $ref: '#/components/parameters/showAll'
      responses:
        200:
          description: Successfully get objects
          content:
            application/json:
              schema:
                type: array
                description: Found short video comment objects
                items:
                  $ref: '#/components/schemas/ShortVideoComment'
        default:
          $ref: '#/components/responses/Error'
  /short-video/comment/{commentId}:
    put:
      tags:
        - Short Video
      summary: Updates a short video comment
      operationId: updateComment
      parameters:
        - $ref: '#/components/parameters/commentId'
      requestBody:
        description: Comment object that needs to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShortVideoComment'
        required: true
      responses:
        200:
          description: Short video successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortVideoComment'
        404:
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags:
        - Short Video
      summary: Deletes a short video comment
      operationId: removeComment
      parameters:
        - $ref: '#/components/parameters/commentId'
        - $ref: '#/components/parameters/videoUuid'
      responses:
        204:
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortVideoStatsResponse'
        404:
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    get:
      tags:
        - Short Video
      summary: Retrieves short video comment by id
      operationId: getComment
      parameters:
          - $ref: '#/components/parameters/commentId'
          - $ref: '#/components/parameters/videoUuid'
      responses:
        200:
          description: Successfully get short video comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortVideoComment'
        default:
          $ref: '#/components/responses/Error'
  /short-video/comment/{commentId}/reply:
    post:
      tags:
        - Short Video
      summary: Replies to the short video comment
      operationId: reply
      parameters:
        - $ref: '#/components/parameters/commentId'
      requestBody:
        description: Reply object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShortVideoComment'
        required: true
      responses:
        201:
          description: Created reply to the short video comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortVideoStatsResponse'
        400:
          $ref: '#/components/responses/InvalidInput'
        default:
          $ref: '#/components/responses/Error'
    get:
      tags:
        - Short Video
      summary: Retrieves short video comment replies
      operationId: getCommentReplies
      x-spring-paginated: true
      parameters:
        - $ref: '#/components/parameters/commentId'
        - $ref: '#/components/parameters/videoUuid'
        - $ref: '#/components/parameters/showAll'
      responses:
        200:
          description: Successfully get objects
          content:
            application/json:
              schema:
                type: array
                description: Found short video comment replies
                items:
                  $ref: '#/components/schemas/ShortVideoComment'
        default:
          $ref: '#/components/responses/Error'
...
components:
  schemas:
    Error:
      required:
        - code
        - message
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
...
    ShortVideoComment:
      required:
        - content
      type: object
      properties:
        id:
          description: Comment id
          type: integer
          format: int64
          example: 150
        authorId:
          description: Author id
          type: integer
          format: int64
          example: 150
        content:
          description: Comment content
          type: string
          example: Some content
        videoId:
          description: Unique short video ID
          type: string
          format: uuid
        replyTo:
          description: Reply to comment id
          type: integer
          format: int64
    ShortVideoStatsResponse:
      type: object
      properties:
        id:
          description: Unique short video ID
          type: string
          format: uuid
        likes:
          description: Likes count
          type: integer
          format: int64
        dislikes:
          description: Dislikes count
          type: integer
          format: int64
        comments:
          description: Comments count
          type: integer
          format: int64
        halves:
          description: Halves count
          type: integer
          format: int64
        views:
          description: Views count
          type: integer
          format: int64
        skips:
          description: Skips count
          type: integer
          format: int64
...
  responses:
    Error:
      description: Error accessing the service consistent with http status code
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InvalidInput:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Unable to find resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
...
    commentId:
      description: Comment id
      name: commentId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    videoUuid:
      description: Short Video UUID
      name: videoId
      in: query
      required: true
      schema:
        type: string
        format: uuid
    showAll:
      description: Indicator to show deleted comments.
      name: showAll
      in: query
      required: false
      schema:
        type: boolean
...